src/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==35.0.0
Django==4.2.15
django-celery-results==2.5.1
django-redis==5.2.0
djangorestframework==3.15.2
idna==3.7
kombu==5.4.0
pillow==10.4.0
prompt-toolkit==3.0.47
psycopg2==2.9.9
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
redis==4.6.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13